# select_statement: Calculate _restrictions->need_filtering() only once