@avikivity 2021-12-05
My part in the improvement was increasing the regression in 0bd139e81c7137bacc7ad60e893fdddc4b61c058 (but I guessed this was coming).